move splash components into their own folder
