+++ lib/components/Onboarding/splash/OnboardingButtons.js
