+++ lib/components/Onboarding/splash/OnboardingIntro.js
