+++ lib/components/Onboarding/splash/OnboardingLogo.js
