+++ lib/containers/SplashContainer.js
@@ -5,10 +5,10 @@ import {
   ScrollView
 } from 'react-native'
 import { LinearGradient } from 'expo'
-import OnboardingLogo from '../components/Onboarding/OnboardingLogo'
-import OnboardingIntro from '../components/Onboarding/OnboardingIntro'
-import OnboardingLogin from '../components/Onboarding/OnboardingLogin'
-import OnboardingButtons from '../components/Onboarding/OnboardingButtons'
+import OnboardingLogo from '../components/Onboarding/splash/OnboardingLogo'
+import OnboardingIntro from '../components/Onboarding/splash/OnboardingIntro'
+import OnboardingLogin from '../components/Onboarding/splash/OnboardingLogin'
+import OnboardingButtons from '../components/Onboarding/splash/OnboardingButtons'
 import { fetchCurrentCountry } from 'circles-mobile/lib/actions/OnboardingActions'
 import { addCognitoUserData, initSignIn, confirmSignIn } from 'circles-mobile/lib/actions/AuthActions'
